Changed SDL_WINDOW_INFO to use int for x, y, width, height (prevents precision loss with SDL_WINDOWPOS_CENTERED) Changed Window constructor signature from float to int parameters Both structs now use C++ style struct Name {} instead of typedef struct {} Name Added SDL_WINDOW_SHOWN to DEFAULT_FLAGS Added default initialization to SDL_FLAGS members

+++ src/classes/UI/Window.cpp
@@ -44,7 +44,7 @@ namespace ADS::UI {
      *
      * @note Window dimensions are automatically scaled based on display DPI
      */
-    Window::Window(std::string title, float x, float y, float width, float height, SDL_FLAGS *flags, ImGuiIO* io): io(io)
+    Window::Window(std::string title, int x, int y, int width, int height, SDL_FLAGS *flags, ImGuiIO* io): io(io)
     {
         this->flags = new SDL_FLAGS();
